+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Date format: DD/MM/YYYY
 
 ## [next] - [##/##/2021]
 
+- Update `RadioButton` design:
+  - Added support for radio button content (`RadioButton.content`)
 - Update `ContentDialog` design.
 - Update `NavigationView` design:
   - **BREAKING:** Acryic is not used anymore. Consequently, `useAcrylic` method was removed.